Add a parseDateTime function that handles dates w/and w/out seconds
This seems to happen in logseq settings if the syncAt is set to
a 00 second time. The next time the UI is displayed, the seconds
wont be shown, and then the time will be set to a value without
seconds. Parsing will fail, and force the entire library to
re-sync.